refactor(mock): move mock package to v2 and update package name
- Move mock files from v2/mock to mock/v2 directory
- Update package name from 'mock' to 'v2' in affected files
